Migrate the helper functions to utils.go
